+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+- Support the code generation for the `hiveType` optional on the `Template` annotation
+- Remove the union name from the `visit` positional argument names
+- Added static methods for creating each union member class from the union.
+
+## 1.0.1
+- More flexible version constraints
+
 ## 1.0.0
 
 - Initial version, created by Stagehand